refactor: Componentでだけ使う処理を内包させた
